requirements.txt
# AI Study Chatbot Requirements
# Real OpenAI integration with RAG capabilities for PDF processing

# Core AI & NLP
openai>=1.3.0
tiktoken>=0.5.0

# RAG & Vector Database
chromadb>=0.4.15
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10

# PDF Processing
PyPDF2>=3.0.1
pdfplumber>=0.9.0
pymupdf>=1.23.0

# Text Processing & Embeddings
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=2.0.0

# Web Interface
streamlit>=1.28.0
streamlit-chat>=0.1.1

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.4.0

# Additional utilities
typing-extensions>=4.8.0
requests>=2.31.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0